+++ WoWDatabaseEditor.Common/WDE.TrinityMySqlDatabase/Models/MySqlCreatureTemplateWrath.cs
@@ -25,7 +25,7 @@ namespace WDE.TrinityMySqlDatabase.Models
         public uint ModelId4 { get; set; }
 
         [Column(Name = "scale")]
-        public uint Scale { get; set; }
+        public float Scale { get; set; }
 
         [Column(Name = "gossip_menu_id")] 
         public uint GossipMenuId { get; set; }
@@ -67,7 +67,7 @@ namespace WDE.TrinityMySqlDatabase.Models
         public uint Entry { get; set; }
         
         [Column(Name = "scale")]
-        public uint Scale { get; set; }
+        public float Scale { get; set; }
 
         [Column(Name = "gossip_menu_id")] 
         public uint GossipMenuId { get; set; }
